@@ -1,18 +1,17 @@
 name: git_hooks
 description: Using dart to prevents bad commit or push (git hooks, pre-commit/precommit, pre-push/prepush, post-merge/postmerge and all that stuff...).It is similar to husky.
-version: 0.1.6
+version: 0.1.7
 homepage: https://github.com/xuzhongpeng/git_hooks
 
-
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
-  path: ^1.7.0
-  yaml: ^2.2.1
-  meta: ^1.2.4
+  path: ^1.8.2
+  yaml: ^3.1.1
+  meta: ^1.8.0
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.3.0
+  very_good_analysis: ^3.1.0
+  test: ^1.21.6
 executables:
   git_hooks:
