@@ -1,6 +1,7 @@
-import 'package:path/path.dart' as path;
 import 'dart:io';
+
 import 'package:meta/meta.dart';
+import 'package:path/path.dart' as path;
 
 /// return bool function
 typedef UserBackFun = Future<bool> Function();
@@ -14,19 +15,19 @@ class Utils {
   }
 
   /// get path of git_hooks library
-  static String getOwnPath() {
-    var pacPath = path.fromUri(path.current + '/.packages');
-    var pac = File(pacPath);
-    var a = pac.readAsStringSync();
-    var b = a.split('\n');
-    String resPath;
-    b.forEach((v) {
+  static String? getOwnPath() {
+    final pacPath = path.fromUri('${path.current}/.packages');
+    final pac = File(pacPath);
+    final a = pac.readAsStringSync();
+    final b = a.split('\n');
+    late String resPath;
+    for (final v in b) {
       if (v.startsWith('git_hooks:')) {
-        var index = v.indexOf(':');
-        var lastIndex = v.lastIndexOf('lib');
+        final index = v.indexOf(':');
+        final lastIndex = v.lastIndexOf('lib');
         resPath = v.substring(index + 1, lastIndex);
       }
-    });
+    }
     resPath = path.fromUri(resPath);
     if (path.isRelative(resPath)) {
       resPath = path.canonicalize(resPath);
@@ -39,13 +40,13 @@ class Utils {
 
   /// get commit edit msg from '.git/COMMIT_EDITMSG'
   static String getCommitEditMsg() {
-    var rootDir = Directory.current;
-    var myFile = File(Utils.uri('${rootDir.path}/.git/COMMIT_EDITMSG'));
-    var commitMsg = myFile.readAsStringSync();
+    final rootDir = Directory.current;
+    final myFile = File(Utils.uri('${rootDir.path}/.git/COMMIT_EDITMSG'));
+    final commitMsg = myFile.readAsStringSync();
     return commitMsg;
   }
 
-  static String _gitHookFolder = _rootDir + '/.git/hooks/';
+  static String _gitHookFolder = '$_rootDir/.git/hooks/';
 
   /// get git hooks folder
   static String get gitHookFolder => uri(_gitHookFolder);
