@@ -1,17 +1,17 @@
 import 'dart:io';
+
+import 'package:git_hooks/install/create_hooks.dart';
 import 'package:git_hooks/runtime/git_hooks.dart';
 import 'package:git_hooks/utils/utils.dart';
 import 'package:yaml/yaml.dart';
 
-import 'package:git_hooks/install/create_hooks.dart';
-
 void main(List<String> arguments) {
   if (arguments.isNotEmpty) {
-    var str = arguments[0];
-    if (arguments != null && arguments.isNotEmpty) {
+    final str = arguments[0];
+    if (arguments.isNotEmpty) {
       if (str == 'create') {
         //init files
-        var targetPath;
+        String? targetPath;
         if (arguments.length == 2) {
           targetPath = arguments.last;
         }
@@ -23,21 +23,22 @@ void main(List<String> arguments) {
       } else if (str == '-h' || str == '-help') {
         help();
       } else if (str == '-v' || str == '--version') {
-        var f = File(Utils.uri(Utils.getOwnPath() + '/pubspec.yaml'));
-        var text = f.readAsStringSync();
-        Map yaml = loadYaml(text);
-        String version = yaml['version'];
+        final f = File(Utils.uri('${Utils.getOwnPath()}/pubspec.yaml'));
+        final text = f.readAsStringSync();
+        final yaml = loadYaml(text) as Map<String, String>;
+        final version = yaml['version']!;
         print(version);
       } else if (str == 'uninstall') {
         GitHooks.unInstall();
       } else {
-        print('${str} is not a git_hooks command,see follow');
+        print('$str is not a git_hooks command,see follow');
         print('');
         help();
       }
     } else {
       print(
-          'Too many positional arguments: 1 expected, but ${arguments.length} found');
+        'Too many positional arguments: 1 expected, but ${arguments.length} found',
+      );
       print('');
       help();
     }
@@ -52,6 +53,6 @@ void help() {
   print('Common commands:');
   print('');
   print(' git_hooks create {{targetPath}}');
-  print('   Create hooks files in \'.git/hooks\'');
+  print("   Create hooks files in '.git/hooks'");
   print('');
 }
