@@ -1,49 +1,44 @@
+// ignore_for_file: avoid_print
+
 import 'dart:io';
+
+import 'package:git_hooks/git_hooks.dart';
+import 'package:git_hooks/install/hook_template.dart';
 import 'package:git_hooks/utils/logging.dart';
-import 'package:git_hooks/utils/utils.dart';
-import '../git_hooks.dart';
-import './hook_template.dart';
 import 'package:path/path.dart';
 
-typedef _HooksCommandFile = Future<bool> Function(File file);
 String _rootDir = Directory.current.path;
 
 /// install hooks
 class CreateHooks {
   /// Create files to `.git/hooks` and [targetPath]
-  static Future<bool> copyFile({String targetPath}) async {
-    if (targetPath == null) {
-      targetPath = 'git_hooks.dart';
-    } else {
-      if (!targetPath.endsWith('.dart')) {
-        print('the file what you want to create is not a dart file');
-        exit(1);
-      }
+  static Future<bool> copyFile({String targetPath = 'git_hooks.dart'}) async {
+    if (!targetPath.endsWith('.dart')) {
+      print('the file what you want to create is not a dart file');
+      exit(1);
     }
-    var relativePath = '${_rootDir}/${targetPath}';
-    var hookFile = File(Utils.uri(absolute(_rootDir, relativePath)));
-    var logger = Logger.standard();
+    final relativePath = '$_rootDir/$targetPath';
+    final hookFile = File(Utils.uri(absolute(_rootDir, relativePath)));
+    final logger = Logger.standard();
     try {
       var commonStr = commonHook(Utils.uri(targetPath));
       commonStr = createHeader() + commonStr;
-      var progress = logger.progress('create files');
+      final progress = logger.progress('create files');
       await _hooksCommand((hookFile) async {
         if (!hookFile.existsSync()) {
           await hookFile.create(recursive: true);
         }
         await hookFile.writeAsString(commonStr);
         if (!Platform.isWindows) {
-          await Process.run('chmod', ['777', hookFile.path])
-              .catchError((onError) {
-            print(onError);
-          });
+          await Process.run('chmod', ['777', hookFile.path]).catchError(print);
         }
         return true;
       });
       if (!hookFile.existsSync()) {
-        var exampleStr = userHooks;
-        hookFile.createSync(recursive: true);
-        hookFile.writeAsStringSync(exampleStr);
+        const exampleStr = userHooks;
+        hookFile
+          ..createSync(recursive: true)
+          ..writeAsStringSync(exampleStr);
       }
       print('All files wrote successful!');
       progress.finish(showTiming: true);
@@ -59,24 +54,26 @@ class CreateHooks {
   static Future<String> getTargetFilePath() async {
     var commandPath = '';
     await _hooksCommand((hookFile) async {
-      var hookTemplate = hookFile.readAsStringSync();
-      var match =
+      final hookTemplate = hookFile.readAsStringSync();
+      final match =
           RegExp(r'dart\s(\S+)\s\$hookName').allMatches(hookTemplate).first;
-      commandPath = match.group(1);
+      commandPath = match.group(1)!;
       return false;
     });
     return commandPath;
   }
 
-  static Future<void> _hooksCommand(_HooksCommandFile callBack) async {
-    var gitDir = Directory(Utils.uri(_rootDir + '/.git/'));
-    var gitHookDir = Utils.gitHookFolder;
+  static Future<void> _hooksCommand(
+    Future<bool> Function(File) callBack,
+  ) async {
+    final gitDir = Directory(Utils.uri('$_rootDir/.git/'));
+    final gitHookDir = Utils.gitHookFolder;
     if (!gitDir.existsSync()) {
       throw ArgumentError('.git is not exists in your project');
     }
-    for (var hook in hookList.values) {
-      var path = gitHookDir + hook;
-      var hookFile = File(path);
+    for (final hook in hookList.values) {
+      final path = gitHookDir + hook;
+      final hookFile = File(path);
       if (!await callBack(hookFile)) {
         return;
       }
