@@ -1,8 +1,9 @@
 import 'dart:io';
-import 'package:git_hooks/utils/logging.dart';
-import 'package:git_hooks/utils/type.dart';
+
 import 'package:git_hooks/install/create_hooks.dart';
 import 'package:git_hooks/uninstall/deleteFiles.dart';
+import 'package:git_hooks/utils/logging.dart';
+import 'package:git_hooks/utils/type.dart';
 import 'package:git_hooks/utils/utils.dart';
 
 /// create files or call hooks functions
@@ -11,22 +12,25 @@ class GitHooks {
 
   /// create files from dart codes.
   /// [targetPath] is the absolute path
-  static void init({String targetPath}) async {
+  static Future<void> init({required String targetPath}) async {
     await Process.run('git_hooks', ['-v']).catchError((onError) async {
-      var result = await Process.run('pub', [
+      final ownPath = Utils.getOwnPath();
+
+      final result = await Process.run('pub', [
         'global',
         'activate',
         '--source',
         'path',
-        Utils.getOwnPath()
-      ]).catchError((onError) {
-        print(onError);
-      });
+        if (ownPath != null) ownPath,
+      ]).catchError(print);
       print(result.stdout);
       if (result.stderr.length != 0) {
-        print(_ansi.error(result.stderr));
-        print(_ansi.subtle(
-            'You can check \'git_hooks\' in your pubspec.yaml,and use \'pub get\' or \'flutter pub get\' again'));
+        print(_ansi.error(result.stderr.toString()));
+        print(
+          _ansi.subtle(
+            "You can check 'git_hooks' in your pubspec.yaml,and use 'pub get' or 'flutter pub get' again",
+          ),
+        );
         exit(1);
       }
       await CreateHooks.copyFile(targetPath: targetPath);
@@ -34,13 +38,13 @@ class GitHooks {
   }
 
   /// unInstall git_hooks
-  static void unInstall({String path}) async {
+  static Future<void> unInstall() async {
     await deleteFiles();
   }
 
   /// get target file path.
   /// returns the path that the git hooks points to.
-  static Future<String> getTargetFilePath({String path}) async {
+  static Future<String> getTargetFilePath() async {
     return CreateHooks.getTargetFilePath();
   }
 
@@ -52,19 +56,22 @@ class GitHooks {
   /// GitHooks.call(arguments, params);
   /// ```
   /// [argument] is just passthrough from main methods. It may ['pre-commit','commit-msg'] from [hookList]
-  static void call(List<String> argument, Map<Git, UserBackFun> params) async {
-    var type = argument[0];
+  static Future<void> call(
+    List<String> argument,
+    Map<Git, UserBackFun> params,
+  ) async {
+    final type = argument[0];
     try {
-      params.keys.forEach((userType) async {
+      for (final userType in params.keys) {
         if (hookList[userType.toString().split('.')[1]] == type) {
-          if (!await params[userType]()) {
+          if (!await params[userType]!()) {
             exit(1);
           }
         }
-      });
+      }
     } catch (e) {
       print(e);
-      print('git_hooks crashed when call ${type},check your function');
+      print('git_hooks crashed when call $type,check your function');
     }
   }
 }
